webapp/requirements.txt
aniso8601==9.0.1
biopython==1.79
blinker==1.4
click==8.0.1
et-xmlfile==1.1.0
ete3==3.1.2
Flask==2.0.1
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-Sockets==0.2.1
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.1
itsdangerous==2.0.1
jdcal==1.4.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.20.2
olefile==0.44
openpyxl==3.0.7
Pillow==8.3.1
python-dateutil==2.8.2
pytz==2021.1
pyuwsgi==2.0.21
redis==3.5.3
redis-server==5.0.7
six==1.16.0
Werkzeug==2.0.1
